refactor: remove BowlingGame and integrate into Controller
